+++ src/senaite/core/exportimport/genericsetup/structure.py
@@ -56,7 +56,6 @@ SKIP_EXPORT_TYPES = [
     "Attachment",
     "Batch",
     "ReferenceSample",
-    "SupplyOrder",
     "Worksheet",
 ]