+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 2.3.0 (unreleased)
 ------------------
 
+- #2012 Remove stale supply order code
 - #2011 Removed dependency to Products.TextIndexNG3
 - #2010 Fix TypeError in Analyses Listing View